ui/requirements.txt
streamlit==1.39.0
psycopg2-binary==2.9.9
pandas==2.2.2
matplotlib==3.8.4